pyproject.toml
[tool.poetry]
name = "opa-quotes-storage"
version = "0.1.0"
description = "TimescaleDB storage for real-time market quotes"
authors = ["OPA Team <oscar@opa-machine.com>"]
readme = "README.md"
packages = [{include = "opa_quotes_storage", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12,<3.13"  # Fixed by OPA-63
pydantic = "^2.5"
pydantic-settings = "^2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
alembic = "^1.13"
pyyaml = "^6.0"
prometheus-client = "^0.24"
psycopg2-binary = "^2.9"
yfinance = "^0.2"
opa-shared-utils = { git = "https://github.com/Ocaxtar/opa-shared-utils.git", tag = "v0.1.1" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^7.0"
pytest-asyncio = "^0.23"
ruff = "^0.1"
mypy = "^1.7"
httpx = "^0.25"  # For testing

[tool.poetry.scripts]
opa_quotes_storage = "opa_quotes_storage.main:main"

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
